+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## Unreleased
 
+## 0.5.3
+
 - Support dynamic imports
-- Output metafile to dist with build.metafile config option or --meta in the build command
+- Output metafile to dist with build.metafile config option or `--meta` in the build command
 - Fix HMR for JSON imports
-- Add --port <number> to the CLI
+- Add `--port <number>` to the CLI
+- Add `server.qrCode` option to print network URL as QR code (`--qr` in the CLI)
 
 ## 0.5.2